// dep/ileDocs/src/parser/iledocs/Type.ts
export enum Type {
    Integer = "Integer",
    Char = "Char",
    Varchar = "Varchar",
    Binary = "Binary",
    PackedDecimal = "PackedDecimal",
    ZonedDecimal = "ZonedDecimal",
    Float = "Float",
    Pointer = "Pointer",
    UnsignedInteger = "UnsignedInteger",
    Boolean = "Boolean",
    Date = "Date",
    Timestamp = "Timestamp",
    Time = "Time",
    DataStructure = "DataStructure",
    Graph = "Graph",
    Object = "Object",
    UCS2 = "UCS2",
    Vargraph = "Vargraph",
    VarUCS2 = "VarUCS2"
}